Extract alonzo-specific chain-trace tests in a separate module
