+++ eras/alonzo/test-suite/cardano-ledger-alonzo-test.cabal
@@ -76,6 +76,7 @@ test-suite cardano-ledger-alonzo-test
     main-is:          Tests.hs
     hs-source-dirs:   test
     other-modules:
+        Test.Cardano.Ledger.Alonzo.ChainTrace
         Test.Cardano.Ledger.Alonzo.Golden
         Test.Cardano.Ledger.Alonzo.PropertyTests
         Test.Cardano.Ledger.Alonzo.PlutusScriptExamples
